Switch to Cassandra 3.5
Build and mvn -Prun-its are passing.

Several issues found during this upgrade:

1. Thrift Client.describe_keyspace() seems to accept case-sensitive
argument while keyspace in integration tests is created without
quotes and is not found in test. Fixed test to request lowercase
version.

2. Syntax for table creation was changed in cql 3.x so load.cql
scripts were fixed.

3. Cql version 2 is not supported in Cassandra 3.x so default is
changed to 3.4.0.

4. -Dcassandra.storagedir must be specified for Cassandra 3.5 but
directory may contain spaces and should be put in quotes (or escaped).
If whole argment was quoted ("-Dcassandra...=...") then it was treated
as a Main classname during launch (some unknown bug? Ubuntu 14) so
we have to do quotation manually. Also URLEncode of value seems to
work too, it is currently performed on log4j.configuration property
if value contains quotes, because value is converted to URL before
append.